+++ src/ui/terminal/mod.rs
@@ -1,4 +1,1 @@
 pub mod crossterm;
-
-#[cfg(target_family = "unix")]
-pub mod termion;